+++ docs/changelog.md
@@ -4,6 +4,7 @@
 - *Sequencer* - Removed stray debugger, whoops
 - *Effects* - Added `.temporary()`, which causes an effect to not set any flags on any objects, which means a refresh will make the effect disappear
 - *Effects* - Fixed `.zIndex()` not working with `.screenSpace()`
+- *Effects* - Fixed `offset`'s `local` modifier not applying correctly when using `.stretchTo()` and the like
 
 ## Version 3.1.2
 - *Sequencer* - Fixed Sequencer's ready hooks sometimes not firing when canvas was fully ready